feat!: rename cancellable_loop to just cancellable
The previous name, cancellable_loop, was misleading, since the macro didn't actually include any loop. Rename it to clearly indicate it only represents a cancellable operation, which might or might not be a loop (even though it usually is one).

BREAKING CHANGE: changes the name of a publicly exported library macro.